+++ backend/Domain/Interfaces/Repositories/ISessionRepository.cs
@@ -7,7 +7,7 @@ public interface ISessionRepository
   Task<IReadOnlyCollection<Session>> GetAllSessionsAsync(Guid userId);
   Task<IReadOnlyCollection<Session>> GetAllActiveSessionsAsync(Guid userId);
   Task<Session?> GetSessionByRefreshTokenAsync(string refreshToken);
-  Task<Session?> GetSessionByUserAndIdAsync(Guid userId, int sessionId);
-  Task<bool> TryMarkSessionAsUsedAsync(int sessionId);
-  Task SetReplacedByAndRevokedAsync(int oldSessionId, int newSessionId);
+  Task<Session?> GetSessionByUserAndIdAsync(Guid userId, long sessionId);
+  Task<bool> TryMarkSessionAsUsedAsync(long sessionId);
+  Task SetReplacedByAndRevokedAsync(long oldSessionId, long newSessionId);
 }
